Refactor derivative generation; add PTIF derivative generation.
Adds a DerivativesManager to manage creation, change, and deletion of
various derivatives based on the 'content' datastream.  By default
derivatives are generated asynchronously, with a separate Resque job for
each derivative to be generated.
Adds a 'multiresImage' datastream to hold PTIF's, JPEG2000's, or other
multi-resolution images for presentation by an image server.
Closes #170.