Adds `tempfile` method to DatastreamBehavior
Method yields a new (closed) tempfile containing the datastream content
to the block.  The tempfile is automatically removed on exiting the block.
An exception is raised if the datastream is empty.  The default file name
is prefixed by the pid (subbing _ for :) and suffixed by the extension
for the datastream media type.  Use the `:prefix` and `:suffix` options
to override.