Adds index field for "unique ids"
`Ddr::Index::Fields::UNIQUE_ID` indexes as searchable (not stored)
and array of unique ids for the object. This includes the Fedora ID,
the permanent ID (if present), and the permanent ID "suffix"
(if permanent ID present) -- i.e., the part of the ARK after the
institutional part (NAAN), a.k.a. the "name".  See
https://wiki.ucop.edu/display/Curation/ARK.

The purpose is to enable one search for unique identifiers.  In
ActiveFedora (subclasses of `Ddr::Models::Base`), use
`.find_by_unique_id(unique_id)`.